master

+++ main.py
@@ -44,7 +44,8 @@ def downloadbook(book):
 
 def delbook(book):
     dir = book.father
-    os.remove("./bookfiles/" + book.info["title"].replace(": ", "：") + "." + book.info["file_type"])
+    if "file_type" in book.info:
+        os.remove("./bookfiles/" + book.info["title"].replace(": ", "：") + "." + book.info["file_type"])
     dir.remove(book)
 
 def search_online(keyword): # 运行在线搜索爬虫
@@ -56,13 +57,13 @@ if __name__ == '__main__':
     addbook(root)
     root.sort("title")
     downloadbook(root.sons[1])
-    temp = hashtable.search("Introduction to Machine Learning with Python: A Guide for Data Scientists")
+    temp = hashtable.search(root.sons[1].info["title"])
     print(temp.info["authors"])
     delbook(root.sons[0])
-    app = QApplication(sys.argv)
-    MainWindow = QMainWindow()
-    ui = Ui_untitled.Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
+    # app = QApplication(sys.argv)
+    # MainWindow = QMainWindow()
+    # ui = Ui_untitled.Ui_MainWindow()
+    # ui.setupUi(MainWindow)
+    # MainWindow.show()
+    # sys.exit(app.exec_())
     __finish__()
